Jira이슈 없음 [Update] 댓글 삭제 확인창 추가

+++ frontend/src/components/CommentListItem.vue
@@ -143,10 +143,12 @@ export default {
     },
 
     deleteComment() {
-      commentStore.actions.deleteComment(this.commentId);
-      const commentTemplate = document.getElementById("comment_template" + this.commentId);
-      commentTemplate.classList.add("d-none");
-      commentTemplate.classList.remove("d-block");
+      if (confirm("댓글을 삭제하시겠습니까?")) {
+        commentStore.actions.deleteComment(this.commentId);
+        const commentTemplate = document.getElementById("comment_template" + this.commentId);
+        commentTemplate.classList.add("d-none");
+        commentTemplate.classList.remove("d-block");
+      }
     },
   },
 };
